So many comments
I added a bunch of decription to the code in cased we get confused about
what we have done or trying to do. I also deleted a lot of useless
stuff. The code looks much cleaner with better documentation.

Ghost-Player collision works. We have to make a GAMEOVER state but as of
right now this collision stops the player.